test: Move tests for RT0 with gravity to main RT0 test module.
